Add files via upload
This is a copy of the code used for my DSC43C8 - Big Data Homework 2 assignment. In this assignment, I used multiple DNN models on CIFAR-10 and MINIST data. In addition to using DNN, the models were run under both ADAM and SGD optimizers. I also used PCA, t-SNE, and UMAP to perform the analysis.